Revert: Set deployment to active for feature teardown as not seeing desired results
Reverting this. Whilst I see that curl command results show that a status for an id is made inactive, it's not making the 'feature' environment inactive. That needs more time to investigate.
It looks like new prs raised for feature make the environment inactive if there is no deploy. The issue is that if a pr with a deploy has resulted in teardown following merge, the feature environment is still shows as active